parser-backend/requirements.txt
filetype==1.2.0
Flask==2.3.2
Flask_Cors==3.0.10
nltk==3.8.1
pymongo==4.3.3
PyPDF2==3.0.1
python-dotenv==1.0.0
